+++ OpenSystemDynamics/src/codemirror/addon/hint/show-hint.js
@@ -244,6 +244,7 @@
     }
     hints.style.left = (left - offsetLeft) + "px";
     hints.style.top = (top - offsetTop) + "px";
+    hints.style.zIndex = "3333"
 
     // If we're at the edge of the screen, then we want the menu to appear on the left of the cursor.
     var winW = parentWindow.innerWidth || Math.max(ownerDocument.body.offsetWidth, ownerDocument.documentElement.offsetWidth);